Django-Middleware-0x03/requirements.txt
Django>=5.2.1
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.3.1
django-filter>=23.5 